dom/workers/test/test_newError.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<html>
<head>
  <title>Test for DOM Worker Threads</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">

  var worker = new Worker("newError_worker.js");

  worker.onmessage = function(event) {
    ok(false, "Shouldn't get a message!");
    SimpleTest.finish();
  }

  worker.onerror = function(event) {
    is(event.message, "Error: foo!", "Got wrong error message!");
    event.preventDefault();
    SimpleTest.finish();
  }

  SimpleTest.waitForExplicitFinish();

</script>
</pre>
</body>
</html>
